Refactored, split classes into separate files, packaged into appropriate packages.
