+++ src/main/kotlin/data/DebugType.kt
@@ -1,0 +1,8 @@
+package data
+
+/**
+ * Created by Mihael Valentin Berčič
+ * on 24/10/2021 at 00:49
+ * using IntelliJ IDEA
+ */
+enum class DebugType { ALL, DEBUG, INFO, ERROR, TRACE, CHAIN, CONSENSUS }
