+++ src/main/kotlin/data/chain/ChainTask.kt
@@ -1,0 +1,11 @@
+package data.chain
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Created by Mihael Valentin Berčič
+ * on 24/10/2021 at 00:46
+ * using IntelliJ IDEA
+ */
+@Serializable
+data class ChainTask(val myTask: SlotDuty, val blockProducer: String, val committee: List<String> = emptyList())
