+++ src/main/kotlin/data/chain/SlotDuty.kt
@@ -1,0 +1,8 @@
+package data.chain
+
+/**
+ * Created by Mihael Valentin Berčič
+ * on 24/10/2021 at 00:49
+ * using IntelliJ IDEA
+ */
+enum class SlotDuty { PRODUCER, COMMITTEE, VALIDATOR }
