+++ src/main/kotlin/data/docker/ContainerMigration.kt
@@ -1,4 +1,4 @@
-package data
+package data.docker
 
 import kotlinx.serialization.Serializable
