+++ src/main/kotlin/data/network/Endpoint.kt
@@ -1,17 +1,10 @@
-package data
+package data.network
 
 /**
  * Created by Mihael Valentin Berčič
  * on 15/10/2020 at 14:44
  * using IntelliJ IDEA
  */
-
-enum class VoteType { FOR, AGAINST, SKIP }
-enum class NetworkRequestType { GET, POST }
-enum class SlotDuty { PRODUCER, COMMITTEE, VALIDATOR }
-enum class DebugType { ALL, DEBUG, INFO, ERROR, TRACE, CHAIN, CONSENSUS }
-
-
 enum class Endpoint(val identification: Byte) {
     Ping(0),
     JoinRequest(1),
